Remove boilerplate not relevant to API server
An attempt at closing #3

I have removed:
 - bower
 - assets directory
 - public css/js
 - HTML templates
 - server routes designating "views"
 - dependencies for any of the above

I have modified:
 - the README
 - server routes to 404 on "/" rather than error with the removal of the
   views

I have updated the dependencies and changed the good-console
configuration to be compliant with the updates. (This is technically
issue #2 but there was little to change with the removal of all of the
above).

The `/messages` routes still work as before, no changes there.